feat: swaggers docs for providers and added more apis and improvements

+++ src/controllers/providerController.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from 'express';
 import catchAsync from '../utils/catchAsync';
 import providerService from '../services/providerService';
+import mongoose from 'mongoose';
 
 const create = catchAsync(async (req: Request, res: Response) => {
   const payload = { ...req.body };
@@ -17,7 +18,7 @@ const create = catchAsync(async (req: Request, res: Response) => {
 const findAll = catchAsync(async (req: Request, res: Response) => {
   const page = Number(req.query.page) || 1;
   const limit = Number(req.query.limit) || 10;
-  const search = (req.query.search as string) || '';
+  const search = typeof req.query.search === 'string' ? req.query.search : '';
   const result = await providerService.findAll(page, limit, search);
 
   res.status(200).json({
@@ -29,6 +30,15 @@ const findAll = catchAsync(async (req: Request, res: Response) => {
 
 const findById = catchAsync(async (req: Request, res: Response) => {
   const { provider_id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(provider_id)) {
+    res.status(400).json({
+      success: false,
+      message: 'Invalid provider_id',
+    });
+    return;
+  }
+
   const result = await providerService.findById(provider_id);
 
   res.status(200).json({
@@ -42,6 +52,13 @@ const updateById = catchAsync(async (req: Request, res: Response) => {
   const { provider_id } = req.params;
   const payload = { ...req.body };
 
+  if (!mongoose.Types.ObjectId.isValid(provider_id)) {
+    res.status(400).json({
+      success: false,
+      message: 'Invalid provider_id',
+    });
+    return;
+  }
   const result = await providerService.updateById(provider_id, payload);
 
   res.status(200).json({
@@ -54,6 +71,13 @@ const updateById = catchAsync(async (req: Request, res: Response) => {
 const deleteById = catchAsync(async (req: Request, res: Response) => {
   const { provider_id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(provider_id)) {
+    res.status(400).json({
+      success: false,
+      message: 'Invalid provider_id',
+    });
+    return;
+  }
   await providerService.deleteById(provider_id);
 
   res.status(200).json({
